Refactor video formatter
