+++ server/controllers/feeds/shared/video-feed-utils.ts
@@ -2,7 +2,7 @@ import { mdToOneLinePlainText, toSafeHtml } from '@server/helpers/markdown'
 import { CONFIG } from '@server/initializers/config'
 import { WEBSERVER } from '@server/initializers/constants'
 import { getServerActor } from '@server/models/application/application'
-import { getCategoryLabel } from '@server/models/video/formatter/video-format-utils'
+import { getCategoryLabel } from '@server/models/video/formatter'
 import { DisplayOnlyForFollowerOptions } from '@server/models/video/sql/video'
 import { VideoModel } from '@server/models/video/video'
 import { MThumbnail, MUserDefault } from '@server/types/models'
