+++ server/models/video/formatter/index.ts
@@ -1,0 +1,2 @@
+export * from './video-activity-pub-format'
+export * from './video-api-format'
