+++ server/models/video/formatter/shared/video-format-utils.ts
@@ -1,0 +1,7 @@
+import { MVideoFile } from '@server/types/models'
+
+export function sortByResolutionDesc (fileA: MVideoFile, fileB: MVideoFile) {
+  if (fileA.resolution < fileB.resolution) return 1
+  if (fileA.resolution === fileB.resolution) return 0
+  return -1
+}
